refactor: Merge testing/auth/custom_auth_mock.pyi, fix spy type
- Merged types/docs from `crudclient/testing/auth/custom_auth_mock.pyi` into `crudclient/testing/auth/custom_auth_mock.py`.

- Implemented missing methods in `CustomAuthMock`.

- Deleted `crudclient/testing/auth/custom_auth_mock.pyi`.

- Added `_record_call` to `EnhancedSpyBase` in `crudclient/testing/spy/enhanced.pyi` to fix test failures.
